[libcontacts] Do not resolve minimized numbers from cache
If we try to resolve a minimized number that we have not previously
queried matches for, then do not resolve to any contact that happens
to be in tha cache that matches that number - there may be a better
match that we should query for.

Once we have queried a specific number, or set the cache to fetch all
numbers, it is safe to resolve from the content of the cache.