[libcontacts] Improve vCard contact import API
This commit improves the SeasideImport API by allowing clients to
specify their own SeasideContactBuilder implementation when
converting the list of Versit documents into storage QContacts.

The SeasideContactBuilder allows the client to parametrise things
like the contact filter used to determine the subset of mergable
contacts, the merge strategy to use, and so on.