[libcontacts] Explicitly tell models to update section bucket index caches. Contributes to JB#45504
When contact data changes in the backend and we have pulled these
changes into the model, we need to tell the attached models that
they need to recalculate their cache of section bucket indexes.

Note that we cannot have a single cache of section bucket indexes
which the models read, as models can have a search filter applied
which means the indexes of contacts in the seaside cache won't
necessarily match what the model expects.