[libcontacts] Update section bucket index cache after deletions. Contributes to JB#46496
When a contact is deleted, we need to recalculate the section bucket
index cache as that contact may have been the only contact in a given
section bucket (display label group).

However, since updating the section bucket index cache is an expensive
operation, we should only do this once per bulk deletion, so this
commit also adds support for bulk deletion operations.